+++ src/bda/plone/cart/browser/cart.js
@@ -2,10 +2,11 @@
 /* global jQuery, bdajax, createCookie, readCookie */
 // Dependencies: jQuery, cookie_functions.js
 
-(function($) {
+(function($, bdajax) {
     "use strict";
 
     var CART_EXECUTION_CONTEXT = null,
+        CART_EXECUTION_URL = null,
         CART_PORTLET_IDENTIFYER = '#portlet-cart',
         CART_VIEWLET_IDENTIFYER = '#cart_viewlet';
 
@@ -14,6 +15,10 @@
         if (execution_context.length) {
             CART_EXECUTION_CONTEXT = execution_context.text();
         }
+        var execution_url = $('#cart');
+        if (execution_url.length) {
+            CART_EXECUTION_URL = execution_url.data('context-url');
+        }
         cart.init();
         cart.query();
         if (window.Faceted !== undefined) {
@@ -21,11 +26,9 @@
                 cart.bind();
             });
         }
-        if (bdajax !== undefined) {
-            $.extend(bdajax.binders, {
-                cart_binder: cart.bind
-            });
-        }
+        $.extend(bdajax.binders, {
+            cart_binder: cart.bind
+        });
     });
 
     function Cart() {
@@ -298,7 +301,7 @@
                 var elem = $(this);
                 var status_message = elem.hasClass('show_status_message');
                 bdajax.request({
-                    url: 'validate_cart_item',
+                    url: CART_EXECUTION_URL + '/validate_cart_item',
                     params: params,
                     type: 'json',
                     success: function(data) {
@@ -360,7 +363,7 @@
                 var elem = $(this);
                 var status_message = elem.hasClass('show_status_message');
                 bdajax.request({
-                    url: 'validate_cart_item',
+                    url: CART_EXECUTION_URL + '/validate_cart_item',
                     params: params,
                     type: 'json',
                     success: function(data) {
@@ -561,7 +564,7 @@
             params.execution_context = CART_EXECUTION_CONTEXT;
         }
         bdajax.request({
-            url: 'cartData',
+            url: CART_EXECUTION_URL + '/cartData',
             params: params,
             type: 'json',
             success: function(data) {
@@ -574,4 +577,4 @@
     var cart = new Cart();
     window.bda_plone_cart = cart;
 
-})(jQuery);
+})(jQuery, bdajax);