+++ src/bda/plone/cart/interfaces.py
@@ -184,9 +184,8 @@ class ICartItemStock(Interface):
     overbook = Attribute(u"Allowed overbooking count. ``None`` "
                          u"means unlimited")
 
-    stock_warning_threshold = Attribute(u"Item stock warning threshold. ``None`` "
-                         u"means unlimited")
-
+    stock_warning_threshold = Attribute(u"Item stock warning threshold. "
+                                        u"``None`` means unlimited")
 
 
 class ICartItemPreviewImage(Interface):